requirements.txt
numpy==1.17.0
h5py==2.9.0
opencv-python==4.1.0.25
tensorflow-gpu==1.14.0
Keras==2.2.5
segmentation-models==0.2.1
scikit-learn==0.21.3
